# Conflicts:
#	news/forms.py
#	templates/base.html
#	users/views.py

+++ users/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.mixins import UserPassesTestMixin
 from django.contrib.auth.views import PasswordChangeView
 from django.core.exceptions import PermissionDenied
 from django.db.models import Q
+from django.template.response import TemplateResponse
 from django.views.generic import CreateView, UpdateView, ListView, DeleteView, FormView
 
 from praksaPlanetSoft.constants import FIRST_PAGE, HTTP_STATUS_400, HTTP_STATUS_401, HTTP_STATUS_200
@@ -80,7 +81,7 @@ class EmployeeCreate(UserPassesTestMixin, CreateView):
         if self.request.user.is_admin or self.request.user.is_superuser:
             return True
         else:
-            raise PermissionDenied("You are not authorized to add new employees")
+            raise PermissionDenied("Niste ovlašteni da dodajete zaposlene")
 
 
 # acts like UpdateView when user is admin or superuser
@@ -116,7 +117,7 @@ class EmployeeDelete(UserPassesTestMixin, DeleteView):
         if self.request.user.is_admin or self.request.user.is_superuser:
             return True
         elif self.request.user.is_authenticated:
-            raise PermissionDenied("You are not authorized to delete employees")
+            raise PermissionDenied("Niste ovlašteni da brišete zaposlene")
         raise
 
 
@@ -139,5 +140,9 @@ class PasswordChange(PasswordChangeView):
 
 @login_required
 def remove_avatar(request):
-    User.objects.filter(id=request.user.id).first().profile_pic.delete(save=True)
-    return HTTP_STATUS_200
+    if request.POST:
+        User.objects.filter(id=request.user.id).first().profile_pic.delete(save=True)
+        return HTTP_STATUS_200
+    else:
+        response = TemplateResponse(request, 'users/employee/avatar-confirm-deletion.html', {})
+        return response
